Move transmissionn target out of channel package and fix visibility of trackEvent API
